services/ai-omni-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
dashscope>=1.20.0
python-dotenv==1.0.1
websockets>=12.0
pydantic>=2.6.1
requests==2.31.0
